Add module-level version of constant-propagation pass.
Summary:
Generally most of our passes should work on module level rather than on a graph
level, even though it might make sense for them to have graph version as well.
Constant propagation is a pretty popular pass that can be useful in many
scenarious as just a cleanup, and for it it definitely makes sense to have a
module-level version.

ghstack-source-id: 29e540a83d4bda92020cb976f7c52d7668c65e3e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/24285